+++ src/main/java/com/solomon/springcrm/Service/CustomerService.java
@@ -7,7 +7,7 @@ public interface CustomerService {
 
   public List<Customer> getCustomers();
 
-  public Customer getCustomer(int index);
+  public Customer getCustomer(long index);
 
   public List<Customer> getCustomer(String firstName, String lastName);
